Restructure Dockerfile and bundle all requirements.
For the time being, this is a kludge. We will merge the dependencies for the assessment
workflow tool and the target app itself. This is, obviously in the real world, and not
ideal, and works because the workflow is in Python and the target system is in Python.
We will have to reconsider this at a later time.